+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_export_policy.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 '''
@@ -42,6 +42,30 @@ options:
     type: str
     description:
       - Name of the vserver to use.
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -101,15 +125,19 @@ class NetAppONTAPExportPolicy():
             from_name=dict(required=False, type='str', default=None),
             vserver=dict(required=True, type='str')
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[('use_lambda', True, ('lambda_config',))]
         )
 
         self.na_helper = NetAppModule()
         self.parameters = self.na_helper.set_parameters(self.module.params)
 
         self.rest_api = netapp_utils.OntapRestAPI(self.module)
+        if self.parameters.get('use_lambda') and not self.rest_api.is_rest():
+            self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
         if self.rest_api.is_rest():
             self.use_rest = True
         elif HAS_NETAPP_LIB is False: