+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/modules/na_ontap_snapshot_policy.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 
-# (c) 2018-2025, NetApp, Inc
+# (c) 2018-2026, NetApp, Inc
 # GNU General Public License v3.0+
 # (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
@@ -90,6 +90,30 @@ options:
     required: false
     type: str
     version_added: 2.9.0
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 EXAMPLES = """
 - name: Create Snapshot policy
@@ -188,10 +212,12 @@ class NetAppOntapSnapshotPolicy(object):
             retention_period=dict(required=False, type="list", elements="str"),
             vserver=dict(required=False, type="str")
         ))
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
             required_if=[
                 ('state', 'present', ['enabled', 'count', 'schedule']),
+                ('use_lambda', True, ['lambda_config']),
             ],
             supports_check_mode=True
         )
@@ -207,6 +233,8 @@ class NetAppOntapSnapshotPolicy(object):
             msg = 'REST requires ONTAP 9.8 or later for snapshot schedules.'
             self.use_rest = self.na_helper.fall_back_to_zapi(self.module, msg, self.parameters)
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             used_unsupported_zapi_properties = [option for option in unsupported_zapi_properties if option in self.parameters]