+++ plugins/modules/na_ontap_export_policy_rule.py
@@ -143,6 +143,31 @@ options:
       - With REST, supported from ONTAP 9.9.1 version.
     type: bool
     version_added: 22.0.0
+
+  lambda_config:
+    description:
+      - Configuration parameters for AWS Lambda proxy functionality.
+      - These option and suboptions are only supported with REST.
+    type: dict
+    version_added: 23.5.0
+    suboptions:
+      function_name:
+        description:
+          - The name of the AWS Lambda function to invoke.
+        type: str
+        required: true
+      aws_region:
+        description:
+          - The name of the AWS region.
+        type: str
+        required: true
+      aws_profile:
+        description:
+          - The name of the AWS profile to use for authentication.
+        type: str
+
+notes:
+  - Supports AWS Lambda proxy functionality when using REST. See the README file for examples.
 '''
 
 EXAMPLES = """
@@ -254,10 +279,11 @@ class NetAppontapExportRule:
             chown_mode=dict(required=False, type='str', choices=['restricted', 'unrestricted']),
             allow_device_creation=dict(required=False, type='bool'),
         ))
-
+        self.argument_spec.update(netapp_utils.na_ontap_lambda_argument_spec())
         self.module = AnsibleModule(
             argument_spec=self.argument_spec,
-            supports_check_mode=True
+            supports_check_mode=True,
+            required_if=[('use_lambda', True, ('lambda_config',))],
         )
 
         self.na_helper = NetAppModule()
@@ -270,6 +296,8 @@ class NetAppontapExportRule:
                                                ['allow_device_creation', (9, 9, 1)], ['chown_mode', (9, 9, 1)]]
         self.use_rest = self.rest_api.is_rest_supported_properties(self.parameters, None, partially_supported_rest_properties)
         if not self.use_rest:
+            if self.parameters.get('use_lambda'):
+                self.module.fail_json(msg="Error: AWS Lambda proxy for ONTAP APIs is only supported with REST.")
             if not netapp_utils.has_netapp_lib():
                 self.module.fail_json(msg=netapp_utils.netapp_lib_is_required())
             self.server = netapp_utils.setup_na_ontap_zapi(module=self.module, vserver=self.parameters['vserver'])